moved the task related stuff to the "shared" module
